Shorten filenames
