+++ Makefile
@@ -1,30 +1,31 @@
 YFLAGS = -dv
 
 OBJS = \
-	bohm.a(lambda_parser.o) \
-	bohm.a(lambda_lexan.o) \
+	bohm.a(parser.o) \
+	bohm.a(lexer.o) \
 	bohm.a(main.o) \
 	bohm.a(copy.o) \
-	bohm.a(sthandler.o) \
-	bohm.a(scope_analysis.o) \
+	bohm.a(symbol.o) \
+	bohm.a(scope.o) \
 	bohm.a(inspect.o) \
-	bohm.a(graphgenerator.o) \
-	bohm.a(destroyer.o) \
+	bohm.a(graph.o) \
+	bohm.a(destroy.o) \
 	bohm.a(loader.o) \
 	bohm.a(reducer.o) \
-	bohm.a(m_stack.o) \
+	bohm.a(stack.o) \
 	bohm.a(readback.o) \
 	bohm.a(garbage.o) \
 	bohm.a(menu.o) \
 	bohm.a(save.o) \
-	bohm.a(crashhandler.o) \
-	bohm.a(errorhandler.o) \
-	bohm.a(numberhandler.o) \
-	bohm.a(dynallhandler.o) \
-	bohm.a(filehandler.o) \
-	bohm.a(stringhandler.o)
+	bohm.a(crash.o) \
+	bohm.a(error.o) \
+	bohm.a(number.o) \
+	bohm.a(dynall.o) \
+	bohm.a(file.o) \
+	bohm.a(string.o)
 
 all: bohm
+	./bohm <examples/tartaglia
 
 debug:
 	$(MAKE) CFLAGS="-g -Wall -Werror"
@@ -36,10 +37,10 @@ bohm.a: $(OBJS)
 
 $(OBJS): bohm.h y.tab.h
 
-lambda_lexan.c: y.tab.h
+lexer.c: y.tab.h
 
-y.tab.h: lambda_parser.c
+y.tab.h: parser.c
 
 clean:
 	-rm -f bohm bohm.a *.o
-	-rm -f lambda_parser.c lambda_lexan.c y.*
+	-rm -f parser.c lexer.c y.*
