+++ inspect.c
@@ -1,6 +1,4 @@
 /****************************************************************/
-/*		          INSPECT.C          			*/
-/****************************************************************/
 /* This module implements the "inspect" directive, that is used */
 /* to visit the graph representing a term.                      */
 /* We plan to integrate the i/o with the YACC specification.    */
