+++ number.c
@@ -1,12 +1,8 @@
 /****************************************************************/
-/*		         DYNALLHANDLER.C			*/
-/****************************************************************/
-/* This module supplies routines for dynamic allocation.	*/
-/* It consists of the following functions:			*/
-/* - malloc_da(): it allocates a given amount of bytes; if it	*/
-/*		  is unable to allocate, it signals this fact;	*/
-/* - strdup_da(): it allocates a given string; if it is unable	*/
-/*		  to allocate, it signals this fact.		*/
+/* This module supplies routines for number handling.		*/
+/* It consists of the following function:			*/
+/* - digits_n(): it computes the number of digits of a given	*/
+/*		 integer.					*/
 /****************************************************************/
 
 
@@ -15,10 +11,6 @@
 /****************************************************************/
 
 #include "bohm.h"
-
-#include <stdio.h>
-#include <string.h>
-#include <malloc.h>
 
 /****************************************************************/
 /* 2. Inclusion of declarations that are being imported.        */
@@ -39,38 +31,19 @@
 /* 5. Definitions of functions to be exported.			*/
 /****************************************************************/
 
- /* The following function implements a control interface for the */
- /* library function malloc(). */
-STRING
-malloc_da(size)
-	unsigned	size;
-					/* size of the object to be */
-					/* allocated */
+ /* The following function computes the number of digits of a given */
+ /* integer. */
+int digits_n(n)
+	int		n;
+					/* integer whose number of */
+					/* digits is to be computed */
 {
-	STRING		p;
+	int		digits;
 
-	p = (STRING)malloc(size);
-	if (p != NULL)
-		return(p);
-	else
-		signal_crash(NOTENOUGHMEMORY);
+	for (digits = 1; (n = n / NUMBASE) != 0; digits++);
+	return(digits);
 }
 
- /* The following function implements a control interface for the */
- /* library function strdup(). */
-STRING
-strdup_da(s)
-	STRING		s;
-					/* string to be allocated */
-{
-	STRING		p;
-
-	p = strdup(s);
-	if (p != NULL)
-		return(p);
-	else
-		signal_crash(NOTENOUGHMEMORY);
-}
 
 /****************************************************************/
 /* 6. Definitions of functions strictly local to the module.	*/
