+++ stack.c
@@ -1,6 +1,4 @@
 /****************************************************************/
-/*		              M_STACK.C                         */
-/****************************************************************/
 /* This module implements the auxiliary stack used during       */
 /* the reduction of a term.                                     */
 /****************************************************************/
